chore: refactor tests to smaller files for readability
